% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_score.R
\name{construct_score}
\alias{construct_score}
\title{Construct a PRS}
\usage{
construct_score(file = NULL, ped = NULL, map = NULL, assoc = NULL, mode)
}
\arguments{
\item{file}{Basename for genetic files (i.e. ${file}.ped ${file}.map). These must be unzipped. If non-standard, please specify seperately.}

\item{mode}{One of c("weighted", "unweighted", "additive"). For more details see \code{details}.}
}
\value{
An S4 score object.
}
\description{
Given input genetic files and association statistics, generate one of a number of polygenic risk scores. See \code{mode}. Tries to guess between plink input and Oxford format.
}
\details{
This is helper code to construct a simple PRS in R. Note that it is currently unoptimized, meaning that it is currently implimented in R and not RcppArmadillo as it will be in the future. 
First it reads in.
}

