% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_o_together.R
\name{estimate_o_together}
\alias{estimate_o_together}
\title{Estimate the optimal score and return it}
\usage{
estimate_o_together(dirs, n_i, r_i, h_i, h_1)
}
\value{
Nothing
}
\description{
Estimate the optimal score and return it
}

