% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_comorbid_from_optimal.R, R/make_comorbid_prs.R
\name{make_comborbid_from_optimal}
\alias{make_comborbid_from_optimal}
\title{Construct a comorbid PRS from a list of oPRS objects.}
\usage{
make_comborbid_from_optimal(oPRS, h_1)

make_comborbid_from_optimal(oPRS, h_1)
}
\arguments{
\item{bfile}{Vector of base file names of the binary file set used for the analysis. See details for construction and instructions.}

\item{assoc}{Vector of file names and paths to the assoc files with the beta and P values used to construct the score.}

\item{p}{A vector of P value thresholds to test each score at.}

\item{pheno}{Vector or file path to phenotype to use for association analysis. This defaults to using the phenotype in the .fam file provided, though you can optionally provide a PLINK style phenotype file.}

\item{mode}{See details. If mode is set to 1, then P value thresholds are estimated seperately for each of the different scores prior to their combination. If mode is set to 2, then P values thresholds are estimated *all together* to create the optimal score. Mode 2 is significantly slower than mode 1 for obvious reasons.}

\item{training}{The algorithm used to train the scores. Defaults to 10 times cross validation.}

\item{bfile}{Vector of base file names of the binary file set used for the analysis. See details for construction and instructions.}

\item{assoc}{Vector of file names and paths to the assoc files with the beta and P values used to construct the score.}

\item{p}{A vector of P value thresholds to test each score at.}

\item{pheno}{Vector or file path to phenotype to use for association analysis. This defaults to using the phenotype in the .fam file provided, though you can optionally provide a PLINK style phenotype file.}

\item{mode}{See details. If mode is set to 1, then P value thresholds are estimated seperately for each of the different scores prior to their combination. If mode is set to 2, then P values thresholds are estimated *all together* to create the optimal score. Mode 2 is significantly slower than mode 1 for obvious reasons.}

\item{training}{The algorithm used to train the scores. Defaults to 10 times cross validation.}
}
\value{
(Currently) A data frame with three columns: FID, IID, and the generated score.

(Currently) A data frame with three columns: FID, IID, and the generated score.
}
\description{
Construct a comorbid PRS from a list of oPRS objects.

Construct a comorbid PRS from a list of oPRS objects.
}

