% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_optimal_comorbid_prs.R
\name{make_optimal_comborbid_prs}
\alias{make_optimal_comborbid_prs}
\title{Construct and return the optimal comorbid PRS based on a list of fixed P value thresholds.}
\usage{
make_optimal_comborbid_prs(bfile, assoc, p, pheno = NULL, n_i, r_i, h_i, h_1,
  mode = 1, training = "cv")
}
\arguments{
\item{bfile}{Vector of base file names of the binary file set used for the analysis. See details for construction and instructions.}

\item{assoc}{Vector of file names and paths to the assoc files with the beta and P values used to construct the score.}

\item{p}{A vector of P value thresholds to test each score at.}

\item{pheno}{Vector or file path to phenotype to use for association analysis. This defaults to using the phenotype in the .fam file provided, though you can optionally provide a PLINK style phenotype file.}

\item{n_i}{Vector of the sample sizes used to deduce the effect sizes presented in assoc.}

\item{r_i}{Vector of the genetic covariance of the comorbid trait with the overall phenotype given in pheno.}

\item{h_i}{Vector of the heritability of each of the comorbid traits.}

\item{h_1}{The chip heritability.}

\item{mode}{See details. If mode is set to 1, then P value thresholds are estimated seperately for each of the different scores prior to their combination. If mode is set to 2, then P values thresholds are estimated *all together* to create the optimal score. Mode 2 is significantly slower than mode 1 for obvious reasons.}

\item{training}{The algorithm used to train the scores. Defaults to 10 times cross validation.}
}
\value{
(Currently) A data frame with three columns: FID, IID, and the generated score.
}
\description{
See our pre-print for an introduction to the method.
}
\details{
All input takes the form of PLINK files. If your input is in a different format, please convert it beforehand. This program requires binary format input for performance reasons. We also recommend clumping SNPs or some other method for removing LD between markers. See the vignette for assistance on this.
}

