% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_optimal_prs.R, R/make_prs.R
\name{make_prs}
\alias{make_prs}
\title{Construct and return the optimal PRS based on either P value or R2.}
\usage{
make_prs(bfile, assoc, p, pheno)

make_prs(bfile, assoc, p, pheno)
}
\arguments{
\item{bfile}{Base file name of the binary file set used for the analysis. See details for construction and instructions.}

\item{assoc}{File name and path to the assoc file with the beta and P values used to construct the score.}

\item{p}{A sequence of P value thresholds at which to cut off the score.}

\item{pheno}{(optional) Vector or file path to phenotype to use for association analysis. This is required by other functions but this function by default just returns the score so you can do the association yourself.}

\item{bfile}{Base file name of the binary file set used for the analysis. See details for construction and instructions.}

\item{assoc}{File name and path to the assoc file with the beta and P values used to construct the score.}

\item{p}{P value threshold to cut off the score.}

\item{pheno}{(optional) Vector or file path to phenotype to use for association analysis. This is required by other functions but this function by default just returns the score so you can do the association yourself.}
}
\value{
(Currently) A data frame with three columns: FID, IID, and the generated score.

(Currently) A data frame with three columns: FID, IID, and the generated score.
}
\description{
See the PRSice paper for an introduction to the method.

To construct an optimal PRS see \code{make_optimal_prs()} and to make a comorbid PRS see \code{make_comorbid_prs()}.
}
\details{
All input takes the form of PLINK files. If your input is in a different format, please convert it beforehand. This program requires binary format input for performance reasons. We also recommend clumping SNPs or some other method for removing LD between markers. See the vignette for assistance on this.

All input takes the form of PLINK files. If your input is in a different format, please convert it beforehand. This program requires binary format input for performance reasons. We also recommend clumping SNPs or some other method for removing LD between markers. See the vignette for assistance on this.
}

