% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_range_file.R
\name{make_range_file}
\alias{make_range_file}
\title{Make a range file for plink}
\usage{
make_range_file(range, file)
}
\arguments{
\item{range}{Specify the range as a vector.}

\item{file}{The output location.}
}
\value{
Nothing
}
\description{
Make a range file for plink
}

