% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_score_p.R
\name{make_score_p}
\alias{make_score_p}
\title{Take an assoc file and make the second param to q-score-range.}
\usage{
make_score_p(assoc, file)
}
\arguments{
\item{assoc}{The file path to the assoc file.}
}
\value{
Nothing.
}
\description{
Take an assoc file and make the second param to q-score-range.
}

