% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_score_p.R, R/make_score_profile.R
\name{make_score_profile}
\alias{make_score_profile}
\title{Take an assoc file and make the second param to q-score-range.}
\usage{
make_score_profile(assoc, file)

make_score_profile(assoc, file)
}
\arguments{
\item{assoc}{The file path to the assoc file.}

\item{assoc}{The file path to the assoc file.}
}
\value{
Nothing.

Nothing.
}
\description{
Take an assoc file and make the second param to q-score-range.

Take an assoc file and make an easy to use .scores file.
}

