% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{new_prs}
\alias{new_prs}
\title{Construct several polygenic risk scores from a matrix of weights.}
\usage{
new_prs(input)
}
\arguments{
\item{name}{Path to .ped file (or binary file, working on this now).}

\item{weights}{A matrix of weights with each row being beta corresponding to the association between SNP at that position and the outcome.}
}
\description{
Construct several polygenic risk scores from a matrix of weights.
}

