% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{plink2r}
\alias{plink2r}
\title{Construct several polygenic risk scores from a matrix of weights.}
\usage{
plink2r(input, debug, n, weights)
}
\arguments{
\item{weights}{A matrix of weights with each row being beta corresponding to the association between SNP at that position and the outcome.}

\item{name}{Path to .ped file (or binary file, working on this now).}
}
\description{
Construct several polygenic risk scores from a matrix of weights.
}

