% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prscore.R
\name{prscore}
\alias{prscore}
\title{prscore: calculate representative score of samples}
\usage{
prscore(dist_matrix, sample_matrix)
}
\arguments{
\item{dist_matrix}{dist matrix generated by VCF2Dis}

\item{sample_matrix}{sample matrix with 2 lines (sampleid and populations)}
}
\value{
score
}
\description{
prscore: calculate representative score of samples
}
\examples{
samples_score<-prscore(dist_matrix,sample_matrix)
}
