% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prselect.R
\name{prselect}
\alias{prselect}
\title{prselect: select n samples as representative samples by prescore}
\usage{
prselect(score, n)
}
\arguments{
\item{score}{the output of prescore}

\item{n}{number of selected samples}
}
\value{
selected n samples in populations
}
\description{
prselect: select n samples as representative samples by prescore
}
\examples{
library(prs)
preselect_matrix<-preselect(score,n)
}
