% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prsplot.R
\name{prsplot}
\alias{prsplot}
\title{Title}
\usage{
prsplot(
  dist_matrix,
  sample_matrix,
  picked_samples,
  type = "pcoa",
  label = TRUE,
  axis1 = "Axis.1",
  axis2 = "Axis.2"
)
}
\arguments{
\item{dist_matrix}{the distance matrix produced by VCF2Dis}

\item{sample_matrix}{sample matrix with sample column and cluster column}

\item{picked_samples}{matrix produced by prselect}

\item{type}{pcoa or representative \link{default: pcoa}}

\item{label}{logical: TRUE or FALSE}

\item{axis1}{charactar Axis.1,Axis.2,Axis.3... \link{default: Axis.1}}

\item{axis2}{charactar Axis.1,Axis.2,Axis.3... \link{default: Axis.2}}
}
\value{
0
}
\description{
Title
}
\examples{
prsplot(dist_matrix,sample_matrix,picked_samples)
}
