% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wprs}
\alias{wprs}
\title{Construct a raw weighted PRS from cleaned input}
\usage{
wprs(map, ped, assoc)
}
\arguments{
\item{map}{plink style .map file path. (char)}

\item{ped}{plink style .ped file path. (char)}

\item{assoc}{A vector of beta coefficients passed from R. (arma::vec)}
}

