% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_score_profile.R
\name{make_score_profile}
\alias{make_score_profile}
\title{Take an assoc file and make an easy to use .scores file.}
\usage{
make_score_profile(assoc, file)
}
\arguments{
\item{assoc}{The file path to the assoc file.}
}
\value{
Nothing.
}
\description{
Take an assoc file and make an easy to use .scores file.
}

